+++ components/src/core/__types__/index.ts
@@ -1,12 +1,40 @@
 import { ComponentType } from 'react';
+import { ImageSourcePropType } from 'react-native';
 
-export type WithTheme<T> = T & {
-    theme: ReactNativePaper.Theme;
+// Icon-Specific Types
+export type IconComponentProps = {
+    /** The size of the icon to render */
+    size?: number;
+
+    /** The color of the icon to render
+     *
+     * Default: Theme.colors.text
+     */
+    color?: string;
+
+    /** If true, the icon should scale with the system font size
+     *
+     * Default: true
+     */
+    allowFontScaling?: boolean;
+};
+export type IconFunctionProps = IconComponentProps & {
+    /** The direction of the current device language */
+    direction?: 'rtl' | 'ltr';
 };
 
+export type IconFamily = {
+    name: string;
+    family?: 'material' | 'pxblue' | 'material-community';
+    direction?: 'rtl' | 'ltr' | 'auto';
+    allowFontScaling?: boolean;
+};
+export type IconSourceBase = string | ImageSourcePropType;
+export type IconSource = IconFamily | IconSourceBase | JSX.Element | ComponentType<IconFunctionProps>;
+
 export type HeaderIcon = {
-    /** Name of the icon */
-    icon: ComponentType<{ size: number; color: string }>;
+    /** Component to render for the icon */
+    icon: IconSource;
 
     /** Callback when icon is pressed */
     onPress?: () => void;